requirements.txt
pyyaml
pycocotools

torch==1.7.1
torchvision==0.8.2
Pillow
tqdm
numpy
opencv-python
opencv-contrib-python
matplotlib
tqdm

git+https://github.com/facebookresearch/detectron2.git@v0.3
